// src/types.ts
export type dynamic = { 
    [x: string] : any 
}

export type ValueOf<T> = T[keyof T];

export type CalendarWeekdayFormat = "long" | "short" | "narrow"
export type CalendarMonthFormat = CalendarWeekdayFormat | "numeric" | "2-digit"

export enum CropShape {
    Circle = "circle",
    Square = "square"
}

export const ANCHOR = {
    TopLeft: "top left",
    TopRight: "top right",
    TopCenter: "top center",
    BottomLeft: "bottom left",
    BottomRight: "bottom right",
} as const;
export const AnchorType = ANCHOR

export enum KeyCode {
    
    Backspace = 8,
    Tab = 9,
    Enter = 13,
    Shift = 16,
    Ctrl = 17,
    Alt = 18,
    PauseBreak = 19,
    Command = 19,
    CapsLock = 20,
    Escape = 27,
    Space = 32,
    PageUp = 33,
    PageDown = 34,
    End = 35,
    Home = 36,
    ArrowLeft = 37,
    ArrowUp = 38,
    ArrowRight = 39,
    ArrowDown = 40,
    Insert = 45,
    Delete = 46,
  
    // Number keys
    Digit0 = 48,
    Digit1 = 49,
    Digit2 = 50,
    Digit3 = 51,
    Digit4 = 52,
    Digit5 = 53,
    Digit6 = 54,
    Digit7 = 55,
    Digit8 = 56,
    Digit9 = 57,
  
    // Letter keys
    KeyA = 65,
    KeyB = 66,
    KeyC = 67,
    KeyD = 68,
    KeyE = 69,
    KeyF = 70,
    KeyG = 71,
    KeyH = 72,
    KeyI = 73,
    KeyJ = 74,
    KeyK = 75,
    KeyL = 76,
    KeyM = 77,
    KeyN = 78,
    KeyO = 79,
    KeyP = 80,
    KeyQ = 81,
    KeyR = 82,
    KeyS = 83,
    KeyT = 84,
    KeyU = 85,
    KeyV = 86,
    KeyW = 87,
    KeyX = 88,
    KeyY = 89,
    KeyZ = 90,
  
    // Numpad keys
    Numpad0 = 96,
    Numpad1 = 97,
    Numpad2 = 98,
    Numpad3 = 99,
    Numpad4 = 100,
    Numpad5 = 101,
    Numpad6 = 102,
    Numpad7 = 103,
    Numpad8 = 104,
    Numpad9 = 105,
    NumpadMultiply = 106,
    NumpadAdd = 107,
    NumpadSubtract = 109,
    NumpadDecimal = 110,
    NumpadDivide = 111,
  
    // Function keys
    F1 = 112,
    F2 = 113,
    F3 = 114,
    F4 = 115,
    F5 = 116,
    F6 = 117,
    F7 = 118,
    F8 = 119,
    F9 = 120,
    F10 = 121,
    F11 = 122,
    F12 = 123,
  
    // Other keys
    NumLock = 144,
    ScrollLock = 145,
    Semicolon = 186,  // ;
    Equal = 187,      // =
    Comma = 188,      // ,
    Minus = 189,      // -
    Period = 190,     // .
    Slash = 191,      // /
    Backquote = 192,  // `
    BracketLeft = 219,  // [
    Backslash = 220,    // \
    BracketRight = 221, // ]
    Quote = 222,        // '
}